// xmlns/www.opengis.net/wms.d.ts
interface BaseType {
	_exists: boolean;
	_namespace: string;
}
interface _AttributionType extends BaseType {
	LogoURL?: LogoURLType;
	/** An OnlineResource is typically an HTTP URL.  The URL is placed in
	  * the xlink:href attribute, and the value "simple" is placed in the
	  * xlink:type attribute. */
	OnlineResource?: OnlineResourceType;
	/** The Title is for informative display to a human. */
	Title?: string;
}
interface AttributionType extends _AttributionType { constructor: { new(): AttributionType }; }

interface _AuthorityURLType extends BaseType {
	name: string;
	/** An OnlineResource is typically an HTTP URL.  The URL is placed in
	  * the xlink:href attribute, and the value "simple" is placed in the
	  * xlink:type attribute. */
	OnlineResource: OnlineResourceType;
}
interface AuthorityURLType extends _AuthorityURLType { constructor: { new(): AuthorityURLType }; }

interface _BoundingBoxType extends BaseType {
	CRS: string;
	maxx: number;
	maxy: number;
	minx: number;
	miny: number;
	resx: number;
	resy: number;
}
interface BoundingBoxType extends _BoundingBoxType { constructor: { new(): BoundingBoxType }; }
// ...
interface DataURLType extends _DataURLType { constructor: { new(): DataURLType }; }
// ...
interface DimensionType extends _DimensionType { constructor: { new(): DimensionType }; }

interface _EX_GeographicBoundingBoxType extends BaseType {
	eastBoundLongitude: number;
	northBoundLatitude: number;
	southBoundLatitude: number;
	westBoundLongitude: number;
}
interface EX_GeographicBoundingBoxType extends _EX_GeographicBoundingBoxType { constructor: { new(): EX_GeographicBoundingBoxType }; }
// ...
interface FeatureListURLType extends _FeatureListURLType { constructor: { new(): FeatureListURLType }; }
// ...
interface IdentifierType extends _IdentifierType { constructor: { new(): IdentifierType }; }

interface _KeywordListType extends BaseType {
	/** A single keyword or phrase. */
	Keyword?: KeywordType[];
}
interface KeywordListType extends _KeywordListType { constructor: { new(): KeywordListType }; }

interface _KeywordType extends Primitive._string {
	vocabulary: string;
}
interface KeywordType extends _KeywordType { constructor: { new(): KeywordType }; }

export type latitudeType = number;
type _latitudeType = Primitive._number;

interface _LayerType extends BaseType {
	cascaded: number;
	fixedHeight: number;
	fixedWidth: number;
	noSubsets: boolean;
	opaque: boolean;
	queryable: boolean;
	/** The abstract is a longer narrative description of an object. */
	Abstract?: string;
	/** Attribution indicates the provider of a Layer or collection of Layers.
	  * The provider's URL, descriptive title string, and/or logo image URL
	  * may be supplied.  Client applications may choose to display one or
	  * more of these items.  A format element indicates the MIME type of
	  * the logo image located at LogoURL.  The logo image's width and height
	  * assist client applications in laying out space to display the logo. */
	Attribution?: AttributionType;
	/** A Map Server may use zero or more Identifier elements to list ID
	  * numbers or labels defined by a particular Authority.  For example,
	  * the Global Change Master Directory (gcmd.gsfc.nasa.gov) defines a
	  * DIF_ID label for every dataset.  The authority name and explanatory
	  * URL are defined in a separate AuthorityURL element, which may be
	  * defined once and inherited by subsidiary layers.  Identifiers
	  * themselves are not inherited. */
	AuthorityURL?: AuthorityURLType[];
	/** The BoundingBox attributes indicate the limits of the bounding box
	  * in units of the specified coordinate reference system. */
	BoundingBox?: BoundingBoxType[];
	/** Identifier for a single Coordinate Reference System (CRS). */
	CRS?: string[];
	/** A Map Server may use DataURL offer a link to the underlying data represented
	  * by a particular layer. */
	DataURL?: DataURLType[];
	/** The Dimension element declares the existence of a dimension and indicates what
	  * values along a dimension are valid. */
	Dimension?: DimensionType[];
	/** The EX_GeographicBoundingBox attributes indicate the limits of the enclosing
	  * rectangle in longitude and latitude decimal degrees. */
	EX_GeographicBoundingBox?: EX_GeographicBoundingBoxType;
	/** A Map Server may use FeatureListURL to point to a list of the
	  * features represented in a Layer. */
	FeatureListURL?: FeatureListURLType[];
	Identifier?: IdentifierType[];
	/** List of keywords or keyword phrases to help catalog searching. */
	KeywordList?: KeywordListType;
	/** Nested list of zero or more map Layers offered by this server. */
	Layer?: LayerType[];
	/** Maximum scale denominator for which it is appropriate to
	  * display this layer. */
	MaxScaleDenominator?: number;
	/** A Map Server may use zero or more MetadataURL elements to offer
	  * detailed, standardized metadata about the data underneath a
	  * particular layer. The type attribute indicates the standard to which
	  * the metadata complies.  The format element indicates how the metadata is structured. */
	MetadataURL?: MetadataURLType[];
	/** Minimum scale denominator for which it is appropriate to
	  * display this layer. */
	MinScaleDenominator?: number;
	/** The Name is typically for machine-to-machine communication. */
	Name?: string;
	/** A Style element lists the name by which a style is requested and a
	  * human-readable title for pick lists, optionally (and ideally)
	  * provides a human-readable description, and optionally gives a style
	  * URL. */
	Style?: StyleType[];
	/** The Title is for informative display to a human. */
	Title: string;
}
interface LayerType extends _LayerType { constructor: { new(): LayerType }; }
// ...
interface MetadataURLType extends _MetadataURLType { constructor: { new(): MetadataURLType }; }
// ...
interface StyleType extends _StyleType { constructor: { new(): StyleType }; }